pyproject.toml
[tool.poetry]
name = "SGBench"
version = "1.0.0"
description = "A collection of metrics for scene graph generation"
authors = ["Julian Lorenz <julian.lorenz@uni-a.de>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering"
]

[tool.poetry.dependencies]
python = "^3.9"
tifffile = { version = "^2024.2.12", optional = true }
imagecodecs = { version = "^2024.1.1", optional = true }
pillow = { version = ">=8.0.0", optional = true }
numpy = "^1.26.4"

[tool.poetry.extras]
all = ["tifffile", "imagecodecs", "pillow"]

[tool.poetry.scripts]
sgbench = "sgbench.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
